refactor(example_crate): nest greeting service module
