+++ Docs/examples/example_crate/src/services/greeting/mod.rs
@@ -1,0 +1,7 @@
+//! Types for creating and using a [`GreetingService`].
+
+mod service;
+mod builder;
+
+pub use builder::GreetingServiceBuilder;
+pub use service::GreetingService;
